break out java-types macros into their own files
makes it easier to find them.

also, improve comments.
